Move card HTML into toggle, add database creation endpoint, clarify HTML formatting arguments
