+++ src/server/ui_endpoints/database.rs
@@ -78,7 +78,6 @@ pub async fn database(
                                     const queryTextarea = document.getElementById('query');
                                     const runButton = document.getElementById('run-query-btn');
                                     runButton.disabled = queryTextarea.value.trim() === '';
-
                                     queryTextarea.addEventListener('input', function() {{
                                         runButton.disabled = this.value.trim() === '';
                                     }});
